Optimize the Create Purchase order and implement receive method in IS sect

+++ src/main/java/com/JK/SIMS/repository/IC_repo/IncomingStock_repository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IncomingStock_repository extends JpaRepository<IncomingStock, Integer> {
+public interface IncomingStock_repository extends JpaRepository<IncomingStock, Long> {
 
     boolean existsByPONumber(String potentialPONumber);
 }
